+++ solution/create element function.js
@@ -1,0 +1,18 @@
+// create Element function
+export default function createElement (tagName, children = [], classes = [], attributes = {}) {
+  const newEl = document.createElement(tagName);
+  children.forEach((kid) => {
+    let child = kid;
+    if (typeof (child) === 'string') {
+      child = document.createTextNode(child);
+    }
+    newEl.append(child);
+  });
+  classes.forEach((cls) => {
+    newEl.classList.add(cls);
+  });
+  Object.entries(attributes).forEach((attr) => {
+    newEl.setAttribute(attr[0], attr[1]);
+  });
+  return newEl;
+}
